# backend/Fms/openapi/index.yaml
openapi: 3.0.1
info:
  title: FMS
  version: v1
servers:
  - url: http://localhost:3333
paths:
...
  '/api/organizations/{organizationId}/users/{userId}':
    put:
      tags:
        - Organization
      summary: Add user to organization
      operationId: organizationAddUser
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    delete:
      tags:
        - Organization
      summary: Remove user from organization
      operationId: organizationRemoveUser
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    get:
      tags:
        - Organization
      summary: Get user role in organization
      operationId: organizationGetUserRole
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationRole'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    post:
      tags:
        - Organization
      summary: Update user role in organization
      operationId: organizationUpdateUserRole
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRole'
          text/json:
            schema:
              $ref: '#/components/schemas/OrganizationRole'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OrganizationRole'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
...
  '/api/users/{id}':
    post:
      tags:
        - User
      summary: Get user profile
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  /api/users:
    post:
      tags:
        - User
      summary: Update user profile
      operationId: updateMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
          text/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
    get:
      tags:
        - User
      summary: Get list of users
      operationId: listUsers
      parameters:
        - name: Query
          in: query
          schema:
            type: string
            example: Jane
          example: Jane
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
...
  '/api/workspaces/{workspaceId}/users/{userId}':
    put:
      tags:
        - Workspace
      summary: Add user to workspace
      operationId: workspaceAddUser
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    delete:
      tags:
        - Workspace
      summary: Remove user from workspace
      operationId: workspaceRemoveUser
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    get:
      tags:
        - Workspace
      summary: Get user role in workspace
      operationId: workspaceGetUserRole
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceRole'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    post:
      tags:
        - Workspace
      summary: Update user role in workspace
      operationId: workspaceUpdateUserRole
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceRole'
          text/json:
            schema:
              $ref: '#/components/schemas/WorkspaceRole'
          application/*+json:
            schema:
              $ref: '#/components/schemas/WorkspaceRole'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
...
components:
  schemas:
...
    OrganizationRole:
      enum:
        - OWNER
        - ADMIN
        - MEMBER
      type: string
...
    PublicClientError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
        validationErrors:
          type: array
          items:
            type: string
          nullable: true
          example:
            - Email can't contain less than 3 characters
            - Password can't be empty
      additionalProperties: false
    PublicError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
      additionalProperties: false
...
    UserListResponse:
      required:
        - items
        - totalCount
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
          example: 165
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
      additionalProperties: false
...
    UserResponse:
      required:
        - firstName
        - id
        - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
        firstName:
          minLength: 1
          type: string
          example: John
        lastName:
          minLength: 1
          type: string
          example: Snow
      additionalProperties: false
...
    UserUpdate:
      type: object
      properties:
        firstName:
          minLength: 3
          type: string
          nullable: true
          example: John
        lastName:
          minLength: 3
          type: string
          nullable: true
          example: Snow
      additionalProperties: false
...
    WorkspaceRole:
      enum:
        - OWNER
        - ADMIN
        - COLLABORATOR
        - VIEWER
      type: string
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme.\n\nEnter 'Bearer' [space] and then your token in the text input below.\nExample: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]